refactor: 함수명 및 변수명을 수정한다.
